pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tartanair"
version = "1.0.4"
authors = [
  { name="Yorai Shaoul", email="yorai@cmu.edu" },
  { name="Wenshan Wang", email="wenshanw@cs.cmu.edu " } 
]
description = "TartanAir"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "colorama >= 0.4.0",
    "kornia",
    "matplotlib >= 3.1.2",
    "networkx >= 2.4",
    "numba >= 0.56",
    "numpy",
    "opencv_contrib_python",
    "Pillow",
    "plyfile",
    "pyquaternion",
    "pytransform3d",
    "PyYAML",
    "scipy",
    "pyyaml",
    "pandas",
    "imageio",
    "einops",
    "tqdm"
    ]

[project.urls]
"Homepage" = "https://tartanair.org"
